Color picker: Redesign, (wip, part 15):
Fix updating cards when a new color hex value was entered:
This didn't called the onColorChanged listener,
and as there is no code for updating the cards
without calling the onColorChanged listener in
this scenario, just use the easy way to fix the cards update
and call the onColorChanged listener, on a side effect,
this will also animate the color transition.

Fix updating cards when a new color was set,
and 'color picker view' was the initial widget.

TO-DO
- Add a methode to update the card adapter