Color picker: Redesign, (wip, part 20):
(General)

Fix border color when widgets/drawables/views isn't on the action bar:
As the border color uses the theme colorControlHighlight,
in most cases (when a custom ripple color was set),
the border color will be colorized using the custom ripple color.
The border color should be a visual border between widgets/drawables/views
and background, (especially when the widgets/drawables/views and the background
are using the same color).
So use the default (not customized) theme colorControlHighlight,
(translucent black on light backgrounds and translucent white on dark backgrounds).

To be consistent, use the (customized) theme colorControlHighlight
as border color by default in the ColorPickerView.
All widgets/drawables/views which are using a border
already have a methode to set the border color.