Color picker: Redesign, (wip, part 47):
Fix text-, icon and ripple effect colors on empty card, when using the light theme:
Don't check if the card color is dark, to apply the right (dark or light)
text-, icon and ripple effect colors. On empty favorites, the card color is '0',
and the ColorUtil methode 'isColorDark(int color)' will return 'true'.
Instead check if the 'cardBackgroundColor' (ColorPickerCardAdapter).