Use RecyclerView for weather cards, (wip, part 5):
- Add forecast weather card for the current day to the RecyclerView

Also only change the cards if the weather data timestamp changed

TO-DO:
WeatherInfo: Add a new timestamp for the weather info update time,
using the timestamp provided by the received weather data,
or change the current timestamp (system time at the moment the WeatherInfo
was created) to the timestamp provided by the received weather data.
So the cards will only be updated, if the received weather data changed.
But this shouldn't be done in any of the 'Use RecyclerView for weather cards'
wip commits