ColorPickerListPreference: Dialog title background color:
Background color:
- Theme default color
- Theme primary color
- Color from current selected list item
  (with color transition animation on selected item change)
- Own color

Text color:
Will be set automatically related to the background color,
light text on dark background and vise versa, except if the background
color is set to the default color, the text color will be set to the
theme default primary text color.

TO-DO:
Reset list selection, when the dialog was dismissed.

Find a better color transition animation duration,
in my opinion 350 ms is well recognizable, but to long,
while 250 ms isn't well recognizable but not to long.
Use the 'new' color transition animation duration for
all color transition animation.